+++ src/demos/Demos.Collisions2D/CollisionScenes/TwoDimensional/CircleCastLine.cs
@@ -6,10 +6,10 @@ using System.Numerics;
 
 namespace Demos.Collisions2D.CollisionScenes.TwoDimensional;
 
-internal sealed class LineCircle() : CollisionScene<LineSegment2D, Circle>(Geometry2D.LineCircle)
+internal sealed class CircleCastLine() : CollisionScene<CircleCast, LineSegment2D>(static (a, b) => Geometry2D.CircleCastLine(a, b))
 {
-	private const float _linePointOffset = 64;
-	private const float _circleOffset = 128;
+	private const float _pointOffset = 64;
+	private const float _linePointOffset = 128;
 
 	public override void Update(float dt)
 	{
@@ -17,19 +17,21 @@ internal sealed class LineCircle() : CollisionScene<LineSegment2D, Circle>(Geome
 
 		float halfTime = TotalTime / 2;
 		float quarterTime = TotalTime / 4;
-		A = new LineSegment2D(
+		float endOffset = MathF.Sin(TotalTime) + 4;
+		A = new CircleCast(
+			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(halfTime) * _linePointOffset, MathF.Sin(halfTime) * _linePointOffset),
+			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(halfTime + endOffset) * _linePointOffset, MathF.Sin(halfTime + endOffset) * _linePointOffset),
+			(MathF.Sin(halfTime) * 0.5f + 0.75f) * _pointOffset);
+		B = new LineSegment2D(
 			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(quarterTime) * _linePointOffset, MathF.Sin(quarterTime) * _linePointOffset),
 			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(TotalTime) * _linePointOffset, MathF.Sin(halfTime) * _linePointOffset));
-		B = new Circle(
-			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(TotalTime) * _circleOffset, MathF.Sin(TotalTime) * _circleOffset),
-			64 + MathF.Sin(TotalTime) * 32);
 	}
 
 	public override void Render()
 	{
 		PositionedDrawList drawList = new(ImGui.GetWindowDrawList(), ImGui.GetCursorScreenPos());
 		drawList.AddBackground(CollisionSceneConstants.Size);
-		drawList.AddLine(A, HasCollision);
-		drawList.AddCircle(B, HasCollision);
+		drawList.AddCircleCast(A, HasCollision);
+		drawList.AddLine(B, HasCollision);
 	}
 }